Rewrite to use tokenstorage, not config
Replace use of ~/.globus.cfg storage with ~/.globus_search.db as a
sqlite tokenstorage database managed via globus-sdk-tokenstorage.

As of v0.2.1, globus-sdk-tokenstorage has a featureful storage mechanism
which can write token data and configs for multiple environments (this
is but one of the ways you can use its `namespace` feature).

Update the dependencies of globus-search-cli to use this and discard all
of our code pertaining to configobj manipulation.
The token read/write interface provided is meant to mesh naturally with
OAuthTokenResponse objects, and lets you write or load token data in a
very similar way to the response object.

Tokens are organized by resource server, so it behooves us to add that
as a constant in the code (in globus_search_cli.config).